refactor(lib): move BSP enum files to BSP.Enum
