+++ src/Lumper.Lib/BSP/Enum/StaticPropVersion.cs
@@ -1,0 +1,17 @@
+namespace Lumper.Lib.BSP.Enum;
+
+public enum StaticPropVersion
+{
+    Unknown = 0,
+    V4 = 40,
+    V5 = 50,
+    V6 = 60,
+    V7s = 70,
+    V7 = 71,
+    V8 = 80,
+    V9 = 90,
+    V10 = 100,
+    V11 = 110,
+    V12 = 120,
+    V13 = 130
+}
