+++ src/Lumper.Lib/BSP/Enum/SurfaceFlag.cs
@@ -1,4 +1,5 @@
 namespace Lumper.Lib.BSP.Enum;
+
 public enum SurfaceFlag
 {
     Light,
